+++ e2e/client/client.go
@@ -195,9 +195,8 @@ func (c *TestClient) EnsureOrganization() {
 
 func (c *TestClient) RegisterAlgo(o *AlgoOptions) *asset.Algo {
 	newAlgo := &asset.NewAlgo{
-		Key:      c.ks.GetKey(o.KeyRef),
-		Name:     "Algo test",
-		Category: o.Category,
+		Key:  c.ks.GetKey(o.KeyRef),
+		Name: "Algo test",
 		Description: &asset.Addressable{
 			Checksum:       "1d55e9c55fa7ad6b6a49ad79da897d58be7ce8b76f92ced4c20f361ba3a0af6e",
 			StorageAddress: "http://somewhere.local/desc/" + uuid.NewString(),